feat: v110c support

+++ archives/th19seed/src/main.rs
@@ -7,8 +7,8 @@ use junowen_lib::Th19;
 fn main() -> Result<()> {
     let mut args = args();
     args.next();
-    let seed1 = args.next().and_then(|x| x.parse::<u16>().ok());
-    let seed2 = args.next().and_then(|x| x.parse::<u16>().ok());
+    let seed1 = args.next().and_then(|x| x.parse::<u32>().ok());
+    let seed2 = args.next().and_then(|x| x.parse::<u32>().ok());
     let mut th19 = Th19::new_external_process("th19.exe")?;
     if let (Some(seed1), Some(seed2)) = (seed1, seed2) {
         th19.set_rand_seed1(seed1)?;
